third_party/blink/web_tests/http/tests/websocket/close-event-expected.txt

Make sure WebSocket fires CloseEvent when closed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

WebSocket is open
Received: 'Hello from Simple WSH.'
Closed
PASS closeEventType is "close"
PASS 'wasClean' in closeEvent is true
PASS closeEvent.wasClean is true
PASS Object.getPrototypeOf(closeEvent) === CloseEvent.prototype is true
PASS Object.getPrototypeOf(closeEvent) !== Event.prototype is true
PASS successfullyParsed is true

TEST COMPLETE
